src/main/java/backend/academy/fractal/flame/service/color/chain/impl/OrangeColorChainLinkImpl.java
package backend.academy.fractal.flame.service.color.chain.impl;

import backend.academy.fractal.flame.model.records.ColorRequest;
import backend.academy.fractal.flame.service.color.OrangeTheme;
import backend.academy.fractal.flame.service.color.chain.ColorChainLink;
import java.awt.Color;
import java.util.Optional;
import org.springframework.stereotype.Component;
import static backend.academy.fractal.flame.model.enums.ColorTheme.ORANGE;

@Component
public class OrangeColorChainLinkImpl implements ColorChainLink {

    @Override
    public Optional<Color> getColor(ColorRequest request) {
        if (request.theme().equals(ORANGE)) {
            return Optional.of(OrangeTheme.getRandomColor());
        }
        return Optional.empty();
    }
}
